refactor(app): move mode runners into dedicated modules
